SolidityParam moved to its own file && documented
