+++ lib/solidity/abi.js
@@ -22,9 +22,7 @@
  */
 
 var utils = require('../utils/utils');
-var c = require('../utils/config');
 var coder = require('./coder');
-var f = require('./formatters');
 var solUtils = require('./utils');
 
 /**
