+++ lib/solidity/coder.js
@@ -23,7 +23,7 @@
 var BigNumber = require('bignumber.js');
 var utils = require('../utils/utils');
 var f = require('./formatters');
-var SolidityParam = f.SolidityParam;
+var SolidityParam = require('./param');
 
 var isArrayType = function (type) {
     return type.slice(-2) === '[]';
@@ -43,7 +43,7 @@ SolidityType.prototype.isType = function (name) {
     } else if (this._match === 'prefix') {
         // TODO better type detection!
         return name.indexOf(this._name) === 0;
-    };
+    }
 };
 
 SolidityType.prototype.formatInput = function (param, arrayType) {
@@ -52,9 +52,7 @@ SolidityType.prototype.formatInput = function (param, arrayType) {
         return param.map(function (p) {
             return self._inputFormatter(p);
         }).reduce(function (acc, current) {
-            acc.suffix += current.value;
-            acc.prefix += current.prefix;
-            // TODO: suffix not supported = it's required for nested arrays;
+            acc.appendArrayElement(current);
             return acc;
         }, new SolidityParam('', f.formatInputInt(param.length).value));
     } 
@@ -68,7 +66,7 @@ SolidityType.prototype.formatOutput = function (param, arrayType) {
         var length = new BigNumber(param.prefix, 16);
         for (var i = 0; i < length * 64; i += 64) {
             result.push(this._outputFormatter(new SolidityParam(param.suffix.slice(i, i + 64))));
-        };
+        }
         return result;
     }
     return this._outputFormatter(param);
