+++ lib/solidity/formatters.js
@@ -14,53 +14,17 @@
     You should have received a copy of the GNU Lesser General Public License
     along with ethereum.js.  If not, see <http://www.gnu.org/licenses/>.
 */
-/** @file formatters.js
- * @authors:
- *   Marek Kotewicz <marek@ethdev.com>
+/** 
+ * @file formatters.js
+ * @author Marek Kotewicz <marek@ethdev.com>
  * @date 2015
  */
 
 var BigNumber = require('bignumber.js');
 var utils = require('../utils/utils');
 var c = require('../utils/config');
+var SolidityParam = require('./param');
 
-var SolidityParam = function (value, prefix, suffix) {
-    this.prefix = prefix || '';
-    this.value = value || '';
-    this.suffix = suffix || '';
-};
-
-SolidityParam.prototype.append = function (param) {
-    this.prefix += param.prefix;
-    this.value += param.value;
-    this.suffix += param.suffix;
-};
-
-SolidityParam.prototype.encode = function () {
-    return this.prefix + this.value + this.suffix;
-};
-
-SolidityParam.prototype.shiftValue = function () {
-    var value = this.value.slice(0, 64);
-    this.value = this.value.slice(64);
-    return new SolidityParam(value);
-};
-
-SolidityParam.prototype.shiftBytes = function () {
-    return this.shiftArray(1);   
-};
-
-SolidityParam.prototype.shiftArray = function (length) {
-    var prefix = this.prefix.slice(0, 64);
-    this.prefix = this.value.slice(64);
-    var suffix = this.suffix.slice(0, 64 * length);
-    this.suffix = this.suffix.slice(64 * length);
-    return new SolidityParam('', prefix, suffix);
-};
-
-SolidityParam.prototype.empty = function () {
-    return !this.value.length && !this.prefix.length && !this.suffix.length;
-};
 
 /**
  * Formats input value to byte representation of int
@@ -69,7 +33,7 @@ SolidityParam.prototype.empty = function () {
  *
  * @method formatInputInt
  * @param {String|Number|BigNumber} value that needs to be formatted
- * @returns {String} right-aligned byte representation of int
+ * @returns {SolidityParam}
  */
 var formatInputInt = function (value) {
     var padding = c.ETH_PADDING * 2;
@@ -83,7 +47,7 @@ var formatInputInt = function (value) {
  *
  * @method formatInputString
  * @param {String}
- * @returns {String} left-algined byte representation of string
+ * @returns {SolidityParam}
  */
 var formatInputString = function (value) {
     var result = utils.fromAscii(value, c.ETH_PADDING).substr(2);
@@ -95,7 +59,7 @@ var formatInputString = function (value) {
  *
  * @method formatInputBool
  * @param {Boolean}
- * @returns {String} right-aligned byte representation bool
+ * @returns {SolidityParam}
  */
 var formatInputBool = function (value) {
     var result = '000000000000000000000000000000000000000000000000000000000000000' + (value ?  '1' : '0');
@@ -108,7 +72,7 @@ var formatInputBool = function (value) {
  *
  * @method formatInputReal
  * @param {String|Number|BigNumber}
- * @returns {String} byte representation of real
+ * @returns {SolidityParam}
  */
 var formatInputReal = function (value) {
     return formatInputInt(new BigNumber(value).times(new BigNumber(2).pow(128)));
@@ -129,7 +93,7 @@ var signedIsNegative = function (value) {
  * Formats right-aligned output bytes to int
  *
  * @method formatOutputInt
- * @param {String} bytes
+ * @param {SolidityParam} param
  * @returns {BigNumber} right-aligned output bytes formatted to big number
  */
 var formatOutputInt = function (param) {
@@ -147,7 +111,7 @@ var formatOutputInt = function (param) {
  * Formats right-aligned output bytes to uint
  *
  * @method formatOutputUInt
- * @param {String} bytes
+ * @param {SolidityParam}
  * @returns {BigNumeber} right-aligned output bytes formatted to uint
  */
 var formatOutputUInt = function (param) {
@@ -159,7 +123,7 @@ var formatOutputUInt = function (param) {
  * Formats right-aligned output bytes to real
  *
  * @method formatOutputReal
- * @param {String}
+ * @param {SolidityParam}
  * @returns {BigNumber} input bytes formatted to real
  */
 var formatOutputReal = function (param) {
@@ -170,7 +134,7 @@ var formatOutputReal = function (param) {
  * Formats right-aligned output bytes to ureal
  *
  * @method formatOutputUReal
- * @param {String}
+ * @param {SolidityParam}
  * @returns {BigNumber} input bytes formatted to ureal
  */
 var formatOutputUReal = function (param) {
@@ -181,7 +145,7 @@ var formatOutputUReal = function (param) {
  * Should be used to format output hash
  *
  * @method formatOutputHash
- * @param {String}
+ * @param {SolidityParam}
  * @returns {String} right-aligned output bytes formatted to hex
  */
 var formatOutputHash = function (param) {
@@ -192,7 +156,7 @@ var formatOutputHash = function (param) {
  * Should be used to format output bool
  *
  * @method formatOutputBool
- * @param {String}
+ * @param {SolidityParam}
  * @returns {Boolean} right-aligned input bytes formatted to bool
  */
 var formatOutputBool = function (param) {
@@ -203,7 +167,7 @@ var formatOutputBool = function (param) {
  * Should be used to format output string
  *
  * @method formatOutputString
- * @param {Sttring} left-aligned hex representation of string
+ * @param {SolidityParam} left-aligned hex representation of string
  * @returns {String} ascii string
  */
 var formatOutputString = function (param) {
@@ -215,7 +179,7 @@ var formatOutputString = function (param) {
  * Should be used to format output address
  *
  * @method formatOutputAddress
- * @param {String} right-aligned input bytes
+ * @param {SolidityParam} right-aligned input bytes
  * @returns {String} address
  */
 var formatOutputAddress = function (param) {
@@ -235,7 +199,6 @@ module.exports = {
     formatOutputHash: formatOutputHash,
     formatOutputBool: formatOutputBool,
     formatOutputString: formatOutputString,
-    formatOutputAddress: formatOutputAddress,
-    SolidityParam: SolidityParam
+    formatOutputAddress: formatOutputAddress
 };
